+++ test/NumSharp.UnitTest/Math/NDArray.power.Test.cs
@@ -16,7 +16,7 @@ namespace NumSharp.UnitTest
         {
             var nd = np.arange(3);
             
-            var nd1 = nd.power(2).MakeGeneric<double>();
+            var nd1 = nd.power(2).MakeGeneric<int>();
             
             Assert.IsTrue(nd1[0] == 0);
             Assert.IsTrue(nd1[1] == 1);
